+++ src/platform.ts
@@ -38,7 +38,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     responseType: 'json',
   });
 
-  version = require('../package.json').version; // eslint-disable-line @typescript-eslint/no-var-requires
+  version = require('../package.json').version || '1.12.8'; // eslint-disable-line @typescript-eslint/no-var-requires
   deviceStatus!: deviceResponses;
   registeringDevice!: boolean;
   debugMode!: boolean;
@@ -525,7 +525,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -548,7 +548,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -582,7 +582,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -607,7 +607,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // accessory.context.firmwareRevision = findaccessories.accessoryAttribute.softwareRevision;
@@ -640,7 +640,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -663,7 +663,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -695,7 +695,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -718,7 +718,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -750,7 +750,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -773,7 +773,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -805,7 +805,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -828,7 +828,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -860,7 +860,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -894,7 +894,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -941,7 +941,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -964,7 +964,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -996,7 +996,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1019,7 +1019,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -1051,7 +1051,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1074,7 +1074,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -1106,7 +1106,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.deviceType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
         await this.connectionTypeExistingAccessory(device, existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1129,7 +1129,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.deviceType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `SwitchBot: ${device.deviceType}`;
       await this.connectionTypeNewAccessory(device, accessory);
       // create the accessory handler for the newly create accessory
@@ -1159,7 +1159,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       existingAccessory.context.model = device.remoteType;
       existingAccessory.context.deviceID = device.deviceId;
       existingAccessory.displayName = device.deviceName;
-      existingAccessory.context.firmwareRevision = this.version;
+      existingAccessory.context.firmwareRevision = device.firmware;
       existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeExistingIRAccessory(existingAccessory);
       this.api.updatePlatformAccessories([existingAccessory]);
@@ -1179,7 +1179,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1216,7 +1216,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1239,7 +1239,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1271,7 +1271,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1294,7 +1294,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1326,7 +1326,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1349,7 +1349,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1381,7 +1381,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1404,7 +1404,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1436,7 +1436,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1459,7 +1459,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1491,7 +1491,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1514,7 +1514,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1546,7 +1546,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1569,7 +1569,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory
@@ -1601,7 +1601,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         existingAccessory.context.model = device.remoteType;
         existingAccessory.context.deviceID = device.deviceId;
         existingAccessory.displayName = device.deviceName;
-        existingAccessory.context.firmwareRevision = this.version;
+        existingAccessory.context.firmwareRevision = device.firmware;
         existingAccessory.context.deviceType = `IR: ${device.remoteType}`;
         await this.connectionTypeExistingIRAccessory(existingAccessory);
         this.api.updatePlatformAccessories([existingAccessory]);
@@ -1624,7 +1624,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       accessory.context.device = device;
       accessory.context.model = device.remoteType;
       accessory.context.deviceID = device.deviceId;
-      accessory.context.firmwareRevision = this.version;
+      accessory.context.firmwareRevision = device.firmware;
       accessory.context.deviceType = `IR: ${device.remoteType}`;
       await this.connectionTypeNewIRAccessory(accessory);
       // create the accessory handler for the newly create accessory